requirements.txt
Click==7.0
Flask==1.0.2
flask-csp==0.10
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pyspellchecker==0.4.0
Werkzeug==0.14.1
